backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
httpx
python-dotenv
bson
